boilerplate/mcp-backend/requirements.txt
fastmcp>=2.0.0
azure-identity>=1.15.0
azure-search-documents>=11.4.0
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
httpx>=0.25.0
msal>=1.26.0
